itay cr comments
