+++ crates/ika-types/src/messages_dwallet_mpc.rs
@@ -5,7 +5,8 @@ use crate::dwallet_mpc_error::DwalletMPCError;
 use dwallet_mpc_types::dwallet_mpc::{
     DWalletMPCNetworkKeyScheme, MPCPublicInput, NetworkDecryptionKeyShares,
     DWALLET_MPC_EVENT_STRUCT_NAME, START_DKG_FIRST_ROUND_EVENT_STRUCT_NAME,
-    START_PRESIGN_FIRST_ROUND_EVENT_STRUCT_NAME, START_SIGN_ROUND_EVENT_STRUCT_NAME,
+    START_NETWORK_DKG_EVENT_STRUCT_NAME, START_PRESIGN_FIRST_ROUND_EVENT_STRUCT_NAME,
+    START_SIGN_ROUND_EVENT_STRUCT_NAME,
 };
 use dwallet_mpc_types::dwallet_mpc::{
     MPCMessage, MPCPublicOutput, DWALLET_2PC_MPC_ECDSA_K1_MODULE_NAME, DWALLET_MODULE_NAME,
@@ -44,11 +45,7 @@ pub enum MPCProtocolInitData {
     /// The only round of the network DKG protocol.
     /// Contains the network key scheme, the dWallet network decryption key object ID
     /// and at the end of the session holds the new key version.
-    NetworkDkg(
-        DWalletMPCNetworkKeyScheme,
-        ObjectID,
-        Option<NetworkDecryptionKeyShares>,
-    ),
+    NetworkDkg(DWalletMPCNetworkKeyScheme, StartNetworkDKGEvent),
     /// The round of verifying the encrypted share proof is valid and
     /// that the signature on it is valid.
     /// This is not a real MPC round,
@@ -444,6 +441,27 @@ impl DWalletMPCEventTrait for StartSignEvent {
         StructTag {
             address: *packages_config.ika_system_package_id,
             name: START_SIGN_ROUND_EVENT_STRUCT_NAME.to_owned(),
+            module: DWALLET_MODULE_NAME.to_owned(),
+            type_params: vec![],
+        }
+    }
+}
+
+/// Rust version of the Move [`ika_system::dwallet_network_key::StartNetworkDKGEvent`] type.
+/// It is used to trigger the start of the network DKG process.
+#[derive(Debug, Serialize, Deserialize, Clone, JsonSchema, Eq, PartialEq, Hash)]
+pub struct StartNetworkDKGEvent {
+    pub dwallet_network_decryption_key_id: ID,
+}
+
+impl DWalletMPCEventTrait for StartNetworkDKGEvent {
+    /// This function allows comparing this event with the Move event.
+    /// It is used to detect [`StartNetworkDKGEvent`] events from the chain and initiate the MPC session.
+    /// It is used to trigger the start of the network DKG process.
+    fn type_(packages_config: &IkaPackagesConfig) -> StructTag {
+        StructTag {
+            address: *packages_config.ika_system_package_id,
+            name: START_NETWORK_DKG_EVENT_STRUCT_NAME.to_owned(),
             module: DWALLET_MODULE_NAME.to_owned(),
             type_params: vec![],
         }
